Develop

+++ server/app.js
@@ -14,6 +14,7 @@ app.post('/tokens', (req, res) => {
 
   const { authToken, idFile, config } = req.body
   const figmaStyles = new FigmaTokens(authToken, idFile, config)
+
   figmaStyles.getTokens().then(data => {
     const finalTokens = {}
     Object.entries(data.token).forEach(e => {
@@ -21,7 +22,6 @@ app.post('/tokens', (req, res) => {
       const tokens = { [e[0]]: e[1] }
       fs.writeFileSync(path.resolve(__dirname, `propierties/${e[0]}.json`), JSON.stringify(tokens, null, 2))
     })
-
     res.send(finalTokens)
   }).catch(() => {
     res.send({ type: 'error', message: 'invalid data' })
@@ -32,24 +32,32 @@ app.post('/style-dictionary', async (req, res) => {
   const tokensFiles = fs.readdirSync(path.resolve(__dirname, './propierties')).filter(e => e.includes('.json'))
   await generateConfig(tokensFiles)
   const finalStyles = await readFiles(fs.readdirSync(path.resolve(__dirname, './src/assets/scss/tokens/')))
-  res.send(finalStyles)
+  res.send(finalStyles.map((e, i) => {
+    return { name: tokensFiles[i].split('.')[0], code: e }
+  }))
 })
 
-function deletePropFiles () {
-  const tokensFiles = fs.readdirSync(path.resolve(__dirname, './propierties')).filter(e => e.includes('.json'))
-  tokensFiles.forEach((file, i) => {
-    fs.unlinkSync(path.resolve(__dirname, `./propierties/${file}`))
-  })
+function deletePropFiles() {
+  if (fs.existsSync(path.resolve(__dirname, './propierties'))) {
+    const tokensFiles = fs.readdirSync(path.resolve(__dirname, './propierties')).filter(e => e.includes('.json'))
+    tokensFiles.forEach((file, i) => {
+      fs.unlinkSync(path.resolve(__dirname, `./propierties/${file}`))
+    })
+  }
 
-  const scssFiles = fs.readdirSync(path.resolve(__dirname, './src/assets/scss/tokens'))
-  scssFiles.forEach((file, i) => {
-    fs.unlinkSync(path.resolve(__dirname, `./src/assets/scss/tokens/${file}`))
-  })
+  if (fs.existsSync(path.resolve(__dirname, './src/assets/scss/tokens'))) {
+    const scssFiles = fs.readdirSync(path.resolve(__dirname, './src/assets/scss/tokens'))
+    scssFiles.forEach((file, i) => {
+      fs.unlinkSync(path.resolve(__dirname, `./src/assets/scss/tokens/${file}`))
+    })
+  }
 }
 
-function generateConfig (tokenFiles) {
-  return new Promise((resolve, reject) => {
+function generateConfig(tokenFiles) {
+
+  return new Promise((resolve) => {
     tokenFiles.filter(e => e.includes('.json')).forEach(token => {
+
       const fileName = token.split('.')[0]
       const StyleDictionary = require('style-dictionary').extend({
         source: [`propierties/${fileName}.json`],
@@ -74,16 +82,17 @@ function generateConfig (tokenFiles) {
           return prop.type === 'spacer'
         },
         transformer: function (prop) {
-          return `rem(${prop.original.value})`
+          return `rem(${prop.original.value.replace('px', '')})`
         }
       })
+
       StyleDictionary.buildAllPlatforms()
       resolve()
     })
   })
 }
 
-function readFiles (styleFiles) {
+function readFiles(styleFiles) {
   return new Promise((resolve, reject) => {
     const styles = []
     styleFiles.forEach(fileName => {
